Add switch for shrinkage and prior adj for semi-supervised model.
For semi-supervised version of model, concentration_dist flag in
priors section of configuration file now toggles hierarchical
model over concentrations (shrinkage effect on inferences).

If the shrinkage version is not used, a new prior on $\beta_1$ in
the concentration model is used: $\pi(\beta_1)$ is scaled by
$|\beta_1|^{n_{mis}}$, which yields the UMVUE under a simplified
model. Very successful so far.